pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quench"
version = "0.1.0"
description = "Harden your tensors. Compress everything. A format-aware compression codec for ML tensors."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Bryan", email = "bryan@backwoodsdevelopment.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Archiving :: Compression",
]
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "structlog>=23.1",
]

[project.optional-dependencies]
gpu = ["torch>=2.0"]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "pytest-benchmark>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/itsbryanman/quench"
Repository = "https://github.com/itsbryanman/quench"

[project.scripts]
quench-compress = "quench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
strict = true
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
